Add files via upload
Preliminary config files for Frogger (Sega/Gremlin, Konami pinout) main and sound board. RAM test on main board includes a test of xB300-XB3FF, which appears to be sprite ram. It is behind some buffers so it probably shouldn't be readable by the tester, but it still tests OK on bits 6&7. I've left it in just in case anybody wants to experiment and see why. I can't seem to get inputs to work at all, will need to update those later.